locales: en_CA: update d_fmt [BZ #9842]
The date format in en_CA/LC_TIME specifies the date format as "%d/%m/%y".
However, it should be "%Y-%m-%d".  This is the standard date format in
Canada as specified by the Canadian Standards Association in CSA Z234.5:1989,
which adopts the ISO 8601 standard.

Here's the web page from the National Research Council of Canada
citing ISO 8601 as the standard date/time format in Canada:
http://www.nrc-cnrc.gc.ca/eng/services/time/faq/#Q8
 International Standard ISO 8601 specifies numeric representations
 of date and time. The recommended full format is of the form
 2001-12-31 23:59:28.73 UTC. The intent of this standard is to avoid
 confusion in international communications which can arise with the
 many different national notations. This format has the advantage
 that it permits dates to be readily sorted in chronological order
 by computer systems.

Windows 8+ and OS X also switched to this format.