Add more TCP_* values to netinet/tcp.h.
This patch adds move TCP_* values to sysdeps/gnu/netinet/tcp.h to
bring it up to date with Linux 4.2.  TCP_SAVE_SYN and TCP_SAVED_SYN
are new in 4.2, TCP_NOTSENT_LOWAT and TCP_CC_INFO are older (Szabolcs
previously noted in
<https://sourceware.org/ml/libc-alpha/2015-06/msg00938.html> that this
header was out of date in glibc).

Tested for x86_64 (testsuite, and that installed stripped shared
libraries are unchanged by the patch).

	* sysdeps/gnu/netinet/tcp.h (TCP_NOTSENT_LOWAT): New macro.
	(TCP_CC_INFO): Likewise.
	(TCP_SAVE_SYN): Likewise.
	(TCP_SAVED_SYN): Likewise.