Avoid excess range in results from i386 scalb functions (bug 18981).
i386 scalb / scalbn / scalbln (and thus ldexp) functions for float and
double can return results with excess range (and consequently excess
precision for subnormal results).  As the results of these functions
are fully determined by reference to IEEE 754 operations, this is
unambiguously a bug, apart from the testsuite failures it causes.

This patch makes those functions store their results on the stack and
load them back to eliminate the excess range.  Double rounding is not
a problem, as the only cases where it could occur are when the result
overflows or underflows for extended precision, and then the
double-rounded results are the same as the single-rounded results.

The new macros will be used for more functions, more such macros
added, and existing code refactored to use such macros, in subsequent
patches.

Tested for x86.  Committed.

	[BZ #18981]
	* sysdeps/i386/fpu/i386-math-asm.h: New file.
	* sysdeps/i386/fpu/e_scalb.S: Include <i386-math-asm.h>.
	(__ieee754_scalb): Use DBL_NARROW_EVAL.
	* sysdeps/i386/fpu/e_scalbf.S: Include <i386-math-asm.h>.
	(__ieee754_scalbf): Use FLT_NARROW_EVAL.
	* sysdeps/i386/fpu/s_scalbn.S: Include <i386-math-asm.h>.
	(__scalbn): Use DBL_NARROW_EVAL.
	* sysdeps/i386/fpu/s_scalbnf.S: Include <i386-math-asm.h>.
	(__scalbnf): Use FLT_NARROW_EVAL.