Fix ldbl-128ibm exp10l spurious overflows (bug 16620).
The ldbl-128ibm implementation of exp10l uses a version of log(10)
split into high and low parts - but the low part is negative, so
causing spurious overflows from __ieee754_expl (exp_high) in cases
close to the overflow threshold (I added relevant tests close to the
overflow threshold to the testsuite earlier today).  The same issue
applies close to the underflow threshold as well (except that spurious
underflows in IBM long double arithmetic are harder to fix than the
other deficiencies, so we might end up permitting those for IBM long
double in the libm testsuite, as permitted by ISO C).

This patch fixes it to use a low part rounded downward to 48 bits
instead.  (The choice of 48 instead of 53 bits is to make it more
obviously safe even when the low part of the argument is negative.)

Tested for powerpc.  (Note that because of libgcc bugs with
multiplication very close to LDBL_MAX, libgcc also needs patching for
all the problem cases to be fixed, but this patch is still safe and
correct in the absence of such libgcc fixes.)

	[BZ #16620]
	* sysdeps/ieee754/ldbl-128ibm/e_exp10l.c (log10_high): Use value
	of log (10) rounded downward to 48 bits.
	(log10_low): Use corresponding low part of log (10).